requirements.txt
openai==2.3.0
PyYAML==6.0.3
Requests==2.32.5
websockets==15.0.1
APScheduler==3.11.2